Convert global exposure and isolation to tabular form

Convert global exposure and isolation to tabular form

exposure_isolation_matrix(segregation_results)

Arguments

segregation_results

a segreg object returned by measure_segregation.

Value

a data frame with one row per group and one column per group, where off-diagonal cells are pairwise Exposure indices and diagonal cells are Isolation indices.

Examples

library(sf)
library(segregr)

marilia_sf <- st_read(system.file("extdata/marilia_2010.gpkg",
                                  package = "segregr"))
#> Reading layer `marilia_2010_proj' from data source 
#>   `/home/runner/work/_temp/Library/segregr/extdata/marilia_2010.gpkg' 
#>   using driver `GPKG'
#> Simple feature collection with 295 features and 10 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 597725.2 ymin: 7535553 xmax: 612714.8 ymax: 7551388
#> Projected CRS: SIRGAS 2000 / UTM zone 22S
segregation <- measure_segregation(marilia_sf)
exposure_isolation_matrix(segregation)
#> # A tibble: 9 × 10
#>   group       mw_0_to_05 mw_05_to_1 mw_1_to_2 mw_2_to_3 mw_3_to_5 mw_5_to_10
#>   <fct>            <dbl>      <dbl>     <dbl>     <dbl>     <dbl>      <dbl>
#> 1 mw_0_to_05     0.0552      0.264      0.398     0.140    0.0886     0.0425
#> 2 mw_05_to_1     0.0259      0.236      0.385     0.160    0.112      0.0619
#> 3 mw_1_to_2      0.0208      0.205      0.400     0.170    0.119      0.0651
#> 4 mw_2_to_3      0.0151      0.177      0.354     0.183    0.144      0.0936
#> 5 mw_3_to_5      0.0116      0.149      0.298     0.174    0.173      0.134 
#> 6 mw_5_to_10     0.00784     0.117      0.231     0.159    0.189      0.194 
#> 7 mw_10_to_15    0.00564     0.0912     0.174     0.141    0.200      0.232 
#> 8 mw_15_to_20    0.00470     0.0806     0.157     0.126    0.192      0.246 
#> 9 mw_20_above    0.00470     0.0785     0.154     0.126    0.194      0.241 
#> # ℹ 3 more variables: mw_10_to_15 <dbl>, mw_15_to_20 <dbl>, mw_20_above <dbl>