Collect all local segregation metrics into a spatial sf object

Combines local Dissimilarity, Entropy, H, and pairwise Exposure/Isolation into a single wide sf object, joining results back to the input areal unit geometries.

local_metrics_to_sf(segregation_results, bandwidths = c())

Arguments

segregation_results

a segreg object returned by measure_segregation.

bandwidths

numeric vector. Optional subset of bandwidths to include. When empty (default), all bandwidths are returned.

Value

a spatial sf object with one row per areal unit per bandwidth, containing columns id, bw, dissimilarity, entropy, h, and one column per pairwise exposure/isolation combination.

Examples

library(sf)
library(segregr)

marilia_sf <- st_read(system.file("extdata/marilia_2010.gpkg",
                                  package = "segregr"))
#> Reading layer `marilia_2010_proj' from data source 
#>   `/home/runner/work/_temp/Library/segregr/extdata/marilia_2010.gpkg' 
#>   using driver `GPKG'
#> Simple feature collection with 295 features and 10 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 597725.2 ymin: 7535553 xmax: 612714.8 ymax: 7551388
#> Projected CRS: SIRGAS 2000 / UTM zone 22S
segregation <- measure_segregation(marilia_sf)

local_metrics_to_sf(segregation)
#> Joining with `by = join_by(id, bw)`
#> Simple feature collection with 295 features and 86 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 597725.2 ymin: 7535553 xmax: 612714.8 ymax: 7551388
#> Projected CRS: SIRGAS 2000 / UTM zone 22S
#> First 10 features:
#>                 id bw dissimilarity  entropy             h
#> 1  352900505000001  0  0.0005470273 1.849289 -1.380548e-04
#> 2  352900505000002  0  0.0011001293 1.891925 -1.702489e-04
#> 3  352900505000003  0  0.0017888527 1.850513 -2.179233e-04
#> 4  352900505000004  0  0.0013975371 1.989858 -4.324360e-04
#> 5  352900505000005  0  0.0004308401 1.752560 -4.145553e-05
#> 6  352900505000006  0  0.0004429882 1.600521  3.709038e-04
#> 7  352900505000007  0  0.0002433143 1.743954 -1.725795e-05
#> 8  352900505000008  0  0.0003857752 1.572401  2.994662e-04
#> 9  352900505000009  0  0.0005230570 1.777405 -7.028603e-05
#> 10 352900505000010  0  0.0003233163 1.641489  9.289725e-05
#>    iso_mw_05_to_1_mw_05_to_1 exp_mw_05_to_1_mw_0_to_05
#> 1               3.049899e-04              4.574848e-05
#> 2               3.047101e-05              0.000000e+00
#> 3               3.124129e-05              0.000000e+00
#> 4               1.559238e-04              2.751597e-05
#> 5               8.377819e-04              3.723475e-05
#> 6               1.696003e-03              7.066680e-05
#> 7               5.184872e-04              3.345079e-05
#> 8               6.718253e-04              3.631488e-05
#> 9               2.496099e-04              2.377237e-05
#> 10              1.302486e-04              1.085405e-05
#>    exp_mw_05_to_1_mw_10_to_15 exp_mw_05_to_1_mw_15_to_20
#> 1                1.524949e-05               3.049899e-05
#> 2                5.586351e-05               4.570651e-05
#> 3                4.686193e-05               3.905161e-05
#> 4                7.337592e-05               1.375799e-04
#> 5                3.723475e-05               9.308687e-05
#> 6                4.711120e-05               7.066680e-05
#> 7                6.690158e-05               5.017618e-05
#> 8                3.631488e-05               5.447233e-05
#> 9                9.508948e-05               1.188618e-05
#> 10               1.085405e-05               1.085405e-05
#>    exp_mw_05_to_1_mw_1_to_2 exp_mw_05_to_1_mw_20_above exp_mw_05_to_1_mw_2_to_3
#> 1              4.269858e-04               7.624747e-05             1.982434e-04
#> 2              5.078501e-05               3.554951e-05             6.094201e-05
#> 3              8.200839e-05               4.686193e-05             9.762903e-05
#> 4              2.109558e-04               9.171990e-05             1.467518e-04
#> 5              1.172895e-03               5.585212e-05             8.005471e-04
#> 6              2.332004e-03               0.000000e+00             1.107113e-03
#> 7              8.697205e-04               0.000000e+00             5.017618e-04
#> 8              1.452595e-03               0.000000e+00             4.357786e-04
#> 9              5.348783e-04               1.188618e-05             3.922441e-04
#> 10             3.473296e-04               0.000000e+00             3.039134e-04
#>    exp_mw_05_to_1_mw_3_to_5 exp_mw_05_to_1_mw_5_to_10 exp_mw_0_to_05_mw_05_to_1
#> 1              4.269858e-04              0.0003202394              0.0004669188
#> 2              8.633452e-05              0.0001879045              0.0000000000
#> 3              2.264994e-04              0.0001679219              0.0000000000
#> 4              3.301917e-04              0.0003943956              0.0002808339
#> 5              7.260776e-04              0.0003909649              0.0003800258
#> 6              8.008904e-04              0.0005182232              0.0007212405
#> 7              4.348603e-04              0.0003846841              0.0003414059
#> 8              4.902509e-04              0.0002360467              0.0003706375
#> 9              4.041303e-04              0.0002139513              0.0002426259
#> 10             1.411027e-04              0.0001519567              0.0001107788
#>    iso_mw_0_to_05_mw_0_to_05 exp_mw_0_to_05_mw_10_to_15
#> 1               7.003782e-05               2.334594e-05
#> 2               0.000000e+00               0.000000e+00
#> 3               0.000000e+00               0.000000e+00
#> 4               4.955893e-05               1.321571e-04
#> 5               1.689004e-05               1.689004e-05
#> 6               3.005169e-05               2.003446e-05
#> 7               2.202619e-05               4.405238e-05
#> 8               2.003446e-05               2.003446e-05
#> 9               2.310723e-05               9.242892e-05
#> 10              9.231565e-06               9.231565e-06
#>    exp_mw_0_to_05_mw_15_to_20 exp_mw_0_to_05_mw_1_to_2
#> 1                4.669188e-05             0.0006536863
#> 2                0.000000e+00             0.0000000000
#> 3                0.000000e+00             0.0000000000
#> 4                2.477946e-04             0.0003799518
#> 5                4.222509e-05             0.0005320362
#> 6                3.005169e-05             0.0009917057
#> 7                3.303928e-05             0.0005726809
#> 8                3.005169e-05             0.0008013784
#> 9                1.155361e-05             0.0005199127
#> 10               9.231565e-06             0.0002954101
#>    exp_mw_0_to_05_mw_20_above exp_mw_0_to_05_mw_2_to_3 exp_mw_0_to_05_mw_3_to_5
#> 1                1.167297e-04             0.0003034972             0.0006536863
#> 2                0.000000e+00             0.0000000000             0.0000000000
#> 3                0.000000e+00             0.0000000000             0.0000000000
#> 4                1.651964e-04             0.0002643143             0.0005947071
#> 5                2.533506e-05             0.0003631358             0.0003293557
#> 6                0.000000e+00             0.0004708098             0.0003405858
#> 7                0.000000e+00             0.0003303928             0.0002863405
#> 8                0.000000e+00             0.0002404135             0.0002704652
#> 9                1.155361e-05             0.0003812693             0.0003928229
#> 10               0.000000e+00             0.0002584838             0.0001200103
#>    exp_mw_0_to_05_mw_5_to_10 exp_mw_10_to_15_mw_05_to_1
#> 1               0.0004902647               0.0001580203
#> 2               0.0000000000               0.0005788763
#> 3               0.0000000000               0.0004855990
#> 4               0.0007103446               0.0007603457
#> 5               0.0001773454               0.0003858389
#> 6               0.0002203791               0.0004881819
#> 7               0.0002533012               0.0006932564
#> 8               0.0001302240               0.0003763069
#> 9               0.0002079651               0.0009853488
#> 10              0.0001292419               0.0001124733
#>    exp_mw_10_to_15_mw_0_to_05 iso_mw_10_to_15_mw_10_to_15
#> 1                2.370305e-05                7.901016e-06
#> 2                0.000000e+00                1.061273e-03
#> 3                0.000000e+00                7.283984e-04
#> 4                1.341787e-04                3.578098e-04
#> 5                1.714839e-05                1.714839e-05
#> 6                2.034091e-05                1.356061e-05
#> 7                4.472622e-05                8.945244e-05
#> 8                2.034091e-05                2.034091e-05
#> 9                9.384274e-05                3.753710e-04
#> 10               9.372774e-06                9.372774e-06
#>    exp_mw_10_to_15_mw_15_to_20 exp_mw_10_to_15_mw_1_to_2
#> 1                 1.580203e-05              0.0002212284
#> 2                 8.683144e-04              0.0009647938
#> 3                 6.069987e-04              0.0012746973
#> 4                 6.708933e-04              0.0010287031
#> 5                 4.287098e-05              0.0005401744
#> 6                 2.034091e-05              0.0006712502
#> 7                 6.708933e-05              0.0011628817
#> 8                 3.051137e-05              0.0008136365
#> 9                 4.692137e-05              0.0021114617
#> 10                9.372774e-06              0.0002999288
#>    exp_mw_10_to_15_mw_20_above exp_mw_10_to_15_mw_2_to_3
#> 1                 3.950508e-05              0.0001027132
#> 2                 6.753557e-04              0.0011577526
#> 3                 7.283984e-04              0.0015174967
#> 4                 4.472622e-04              0.0007156195
#> 5                 2.572259e-05              0.0003686905
#> 6                 0.000000e+00              0.0003186743
#> 7                 0.000000e+00              0.0006708933
#> 8                 0.000000e+00              0.0002440910
#> 9                 4.692137e-05              0.0015484053
#> 10                0.000000e+00              0.0002624377
#>    exp_mw_10_to_15_mw_3_to_5 exp_mw_10_to_15_mw_5_to_10
#> 1               0.0002212284               0.0001659213
#> 2               0.0016401495               0.0035697370
#> 3               0.0035205924               0.0026100944
#> 4               0.0016101439               0.0019232274
#> 5               0.0003343937               0.0001800581
#> 6               0.0002305304               0.0001491667
#> 7               0.0005814409               0.0005143515
#> 8               0.0002746023               0.0001322159
#> 9               0.0015953266               0.0008445847
#> 10              0.0001218461               0.0001312188
#>    exp_mw_15_to_20_mw_05_to_1 exp_mw_15_to_20_mw_0_to_05
#> 1                0.0003940149               5.910224e-05
#> 2                0.0005904801               0.000000e+00
#> 3                0.0005045059               0.000000e+00
#> 4                0.0017773874               3.136566e-04
#> 5                0.0012025848               5.344821e-05
#> 6                0.0009129409               3.803921e-05
#> 7                0.0006482237               4.182088e-05
#> 8                0.0007037253               3.803921e-05
#> 9                0.0001535570               1.462448e-05
#> 10               0.0001402230               1.168525e-05
#>    exp_mw_15_to_20_mw_10_to_15 iso_mw_15_to_20_mw_15_to_20
#> 1                 1.970075e-05                3.940149e-05
#> 2                 1.082547e-03                8.857202e-04
#> 3                 7.567588e-04                6.306323e-04
#> 4                 8.364176e-04                1.568283e-03
#> 5                 5.344821e-05                1.336205e-04
#> 6                 2.535947e-05                3.803921e-05
#> 7                 8.364176e-05                6.273132e-05
#> 8                 3.803921e-05                5.705881e-05
#> 9                 5.849792e-05                7.312240e-06
#> 10                1.168525e-05                1.168525e-05
#>    exp_mw_15_to_20_mw_1_to_2 exp_mw_15_to_20_mw_20_above
#> 1               0.0005516209                9.850373e-05
#> 2               0.0009841336                6.888935e-04
#> 3               0.0013243279                7.567588e-04
#> 4               0.0024047007                1.045522e-03
#> 5               0.0016836187                8.017232e-05
#> 6               0.0012552938                0.000000e+00
#> 7               0.0010873429                0.000000e+00
#> 8               0.0015215682                0.000000e+00
#> 9               0.0003290508                7.312240e-06
#> 10              0.0003739279                0.000000e+00
#>    exp_mw_15_to_20_mw_2_to_3 exp_mw_15_to_20_mw_3_to_5
#> 1               0.0002561097              0.0005516209
#> 2               0.0011809603              0.0016730271
#> 3               0.0015765808              0.0036576675
#> 4               0.0016728352              0.0037638793
#> 5               0.0011491365              0.0010422401
#> 6               0.0005959476              0.0004311110
#> 7               0.0006273132              0.0005436715
#> 8               0.0004564705              0.0005135293
#> 9               0.0002413039              0.0002486162
#> 10              0.0003271869              0.0001519082
#>    exp_mw_15_to_20_mw_5_to_10 exp_mw_1_to_2_mw_05_to_1 exp_mw_1_to_2_mw_0_to_05
#> 1                0.0004137157             2.271794e-04             3.407691e-05
#> 2                0.0036412942             2.702036e-05             0.000000e+00
#> 3                0.0027117190             4.363287e-05             0.000000e+00
#> 4                0.0044957447             1.122398e-04             1.980703e-05
#> 5                0.0005612062             6.240430e-04             2.773525e-05
#> 6                0.0002789542             1.240752e-03             5.169799e-05
#> 7                0.0004809401             4.627381e-04             2.985407e-05
#> 8                0.0002472548             7.728589e-04             4.177616e-05
#> 9                0.0001316203             2.845840e-04             2.710324e-05
#> 10               0.0001635934             1.847981e-04             1.539984e-05
#>    exp_mw_1_to_2_mw_10_to_15 exp_mw_1_to_2_mw_15_to_20 iso_mw_1_to_2_mw_1_to_2
#> 1               1.135897e-05              2.271794e-05            3.180512e-04
#> 2               4.953732e-05              4.053054e-05            4.503393e-05
#> 3               6.544931e-05              5.454109e-05            1.145363e-04
#> 4               5.281874e-05              9.903514e-05            1.518539e-04
#> 5               2.773525e-05              6.933811e-05            8.736602e-04
#> 6               3.446533e-05              5.169799e-05            1.706034e-03
#> 7               5.970814e-05              4.478111e-05            7.762058e-04
#> 8               4.177616e-05              6.266423e-05            1.671046e-03
#> 9               1.084130e-04              1.355162e-05            6.098229e-04
#> 10              1.539984e-05              1.539984e-05            4.927948e-04
#>    exp_mw_1_to_2_mw_20_above exp_mw_1_to_2_mw_2_to_3 exp_mw_1_to_2_mw_3_to_5
#> 1               5.679486e-05            1.476666e-04            3.180512e-04
#> 2               3.152375e-05            5.404071e-05            7.655768e-05
#> 3               6.544931e-05            1.363527e-04            3.163383e-04
#> 4               6.602343e-05            1.056375e-04            2.376843e-04
#> 5               4.160287e-05            5.963078e-04            5.408373e-04
#> 6               0.000000e+00            8.099352e-04            5.859106e-04
#> 7               0.000000e+00            4.478111e-04            3.881029e-04
#> 8               0.000000e+00            5.013139e-04            5.639781e-04
#> 9               1.355162e-05            4.472035e-04            4.607551e-04
#> 10              0.000000e+00            4.311955e-04            2.001979e-04
#>    exp_mw_1_to_2_mw_5_to_10 exp_mw_20_above_mw_05_to_1
#> 1              0.0002385384               0.0014224549
#> 2              0.0001666255               0.0006632033
#> 3              0.0002345267               0.0008742453
#> 4              0.0002839007               0.0017111051
#> 5              0.0002912201               0.0010419641
#> 6              0.0003791186               0.0000000000
#> 7              0.0003433218               0.0000000000
#> 8              0.0002715450               0.0000000000
#> 9              0.0002439292               0.0002217459
#> 10             0.0002155977               0.0000000000
#>    exp_mw_20_above_mw_0_to_05 exp_mw_20_above_mw_10_to_15
#> 1                2.133682e-04                7.112274e-05
#> 2                0.000000e+00                1.215873e-03
#> 3                0.000000e+00                1.311368e-03
#> 4                3.019597e-04                8.052259e-04
#> 5                4.630952e-05                4.630952e-05
#> 6                0.000000e+00                0.000000e+00
#> 7                0.000000e+00                0.000000e+00
#> 8                0.000000e+00                0.000000e+00
#> 9                2.111866e-05                8.447462e-05
#> 10               0.000000e+00                0.000000e+00
#>    exp_mw_20_above_mw_15_to_20 exp_mw_20_above_mw_1_to_2
#> 1                 1.422455e-04              0.0019914368
#> 2                 9.948049e-04              0.0011053388
#> 3                 1.092807e-03              0.0022948939
#> 4                 1.509799e-03              0.0023150245
#> 5                 1.157738e-04              0.0014587498
#> 6                 0.000000e+00              0.0000000000
#> 7                 0.000000e+00              0.0000000000
#> 8                 0.000000e+00              0.0000000000
#> 9                 1.055933e-05              0.0004751697
#> 10                0.000000e+00              0.0000000000
#>    iso_mw_20_above_mw_20_above exp_mw_20_above_mw_2_to_3
#> 1                 3.556137e-04              0.0009245957
#> 2                 7.737372e-04              0.0013264065
#> 3                 1.311368e-03              0.0027320165
#> 4                 1.006532e-03              0.0016104518
#> 5                 6.946428e-05              0.0009956546
#> 6                 0.000000e+00              0.0000000000
#> 7                 0.000000e+00              0.0000000000
#> 8                 0.000000e+00              0.0000000000
#> 9                 1.055933e-05              0.0003484578
#> 10                0.000000e+00              0.0000000000
#>    exp_mw_20_above_mw_3_to_5 exp_mw_20_above_mw_5_to_10
#> 1               0.0019914368               0.0014935776
#> 2               0.0018790759               0.0040897535
#> 3               0.0063382783               0.0046990684
#> 4               0.0036235166               0.0043280893
#> 5               0.0009030356               0.0004862499
#> 6               0.0000000000               0.0000000000
#> 7               0.0000000000               0.0000000000
#> 8               0.0000000000               0.0000000000
#> 9               0.0003590171               0.0001900679
#> 10              0.0000000000               0.0000000000
#>    exp_mw_2_to_3_mw_05_to_1 exp_mw_2_to_3_mw_0_to_05 exp_mw_2_to_3_mw_10_to_15
#> 1              2.192613e-04             3.288919e-05              1.096306e-05
#> 2              6.740311e-05             0.000000e+00              1.235724e-04
#> 3              1.079797e-04             0.000000e+00              1.619695e-04
#> 4              1.623105e-04             2.864304e-05              7.638143e-05
#> 5              8.854214e-04             3.935206e-05              3.935206e-05
#> 6              1.224490e-03             5.102041e-05              3.401361e-05
#> 7              5.549588e-04             3.580380e-05              7.160759e-05
#> 8              4.819800e-04             2.605297e-05              2.605297e-05
#> 9              4.338300e-04             4.131714e-05              1.652686e-04
#> 10             3.361345e-04             2.801120e-05              2.801120e-05
#>    exp_mw_2_to_3_mw_15_to_20 exp_mw_2_to_3_mw_1_to_2 exp_mw_2_to_3_mw_20_above
#> 1               2.192613e-05            0.0003069658              5.481531e-05
#> 2               1.011047e-04            0.0001123385              7.863696e-05
#> 3               1.349746e-04            0.0002834467              1.619695e-04
#> 4               1.432152e-04            0.0002195966              9.547679e-05
#> 5               9.838016e-05            0.0012395900              5.902810e-05
#> 6               5.102041e-05            0.0016836735              0.000000e+00
#> 7               5.370569e-05            0.0009308987              0.000000e+00
#> 8               3.907946e-05            0.0010421190              0.000000e+00
#> 9               2.065857e-05            0.0009296357              2.065857e-05
#> 10              2.801120e-05            0.0008963585              0.000000e+00
#>    iso_mw_2_to_3_mw_2_to_3 exp_mw_2_to_3_mw_3_to_5 exp_mw_2_to_3_mw_5_to_10
#> 1             0.0001425198            0.0003069658             0.0002302243
#> 2             0.0001348062            0.0001909755             0.0004156525
#> 3             0.0003374366            0.0007828528             0.0005803909
#> 4             0.0001527629            0.0003437164             0.0004105502
#> 5             0.0008460694            0.0007673652             0.0004131967
#> 6             0.0007993197            0.0005782313             0.0003741497
#> 7             0.0005370569            0.0004654493             0.0004117436
#> 8             0.0003126357            0.0003517152             0.0001693443
#> 9             0.0006817328            0.0007023914             0.0003718543
#> 10            0.0007843137            0.0003641457             0.0003921569
#>    exp_mw_3_to_5_mw_05_to_1 exp_mw_3_to_5_mw_0_to_05 exp_mw_3_to_5_mw_10_to_15
#> 1              0.0005696122             8.544183e-05              2.848061e-05
#> 2              0.0001151729             0.000000e+00              2.111503e-04
#> 3              0.0003021571             0.000000e+00              4.532357e-04
#> 4              0.0004404858             7.773279e-05              2.072874e-04
#> 5              0.0009686099             4.304933e-05              4.304933e-05
#> 6              0.0010684124             4.451718e-05              2.967812e-05
#> 7              0.0005801170             3.742690e-05              7.485380e-05
#> 8              0.0006540098             3.535188e-05              3.535188e-05
#> 9              0.0005391222             5.134497e-05              2.053799e-04
#> 10             0.0001882353             1.568627e-05              1.568627e-05
#>    exp_mw_3_to_5_mw_15_to_20 exp_mw_3_to_5_mw_1_to_2 exp_mw_3_to_5_mw_20_above
#> 1               5.696122e-05            0.0007974571              1.424031e-04
#> 2               1.727594e-04            0.0001919548              1.343684e-04
#> 3               3.776964e-04            0.0007931624              4.532357e-04
#> 4               3.886640e-04            0.0005959514              2.591093e-04
#> 5               1.076233e-04            0.0013560538              6.457399e-05
#> 6               4.451718e-05            0.0014690671              0.000000e+00
#> 7               5.614035e-05            0.0009730994              0.000000e+00
#> 8               5.302782e-05            0.0014140753              0.000000e+00
#> 9               2.567249e-05            0.0011552619              2.567249e-05
#> 10              1.568627e-05            0.0005019608              0.000000e+00
#>    exp_mw_3_to_5_mw_2_to_3 iso_mw_3_to_5_mw_3_to_5 exp_mw_3_to_5_mw_5_to_10
#> 1             0.0003702479            0.0007974571             0.0005980928
#> 2             0.0002303458            0.0003263232             0.0007102329
#> 3             0.0009442409            0.0021906390             0.0016240944
#> 4             0.0004145749            0.0009327935             0.0011141700
#> 5             0.0009255605            0.0008394619             0.0004520179
#> 6             0.0006974359            0.0005045281             0.0003264594
#> 7             0.0005614035            0.0004865497             0.0004304094
#> 8             0.0004242226            0.0004772504             0.0002297872
#> 9             0.0008471921            0.0008728646             0.0004621048
#> 10            0.0004392157            0.0002039216             0.0002196078
#>    exp_mw_5_to_10_mw_05_to_1 exp_mw_5_to_10_mw_0_to_05
#> 1               0.0006032455              9.048682e-05
#> 2               0.0003539620              0.000000e+00
#> 3               0.0003163201              0.000000e+00
#> 4               0.0007429360              1.311064e-04
#> 5               0.0007364734              3.273215e-05
#> 6               0.0009761942              4.067476e-05
#> 7               0.0007246421              4.675110e-05
#> 8               0.0004446491              2.403508e-05
#> 9               0.0004030272              3.838354e-05
#> 10              0.0002862459              2.385382e-05
#>    exp_mw_5_to_10_mw_10_to_15 exp_mw_5_to_10_mw_15_to_20
#> 1                3.016227e-05               6.032455e-05
#> 2                6.489303e-04               5.309430e-04
#> 3                4.744801e-04               3.954001e-04
#> 4                3.496169e-04               6.555318e-04
#> 5                3.273215e-05               8.183038e-05
#> 6                2.711650e-05               4.067476e-05
#> 7                9.350221e-05               7.012665e-05
#> 8                2.403508e-05               3.605263e-05
#> 9                1.535342e-04               1.919177e-05
#> 10               2.385382e-05               2.385382e-05
#>    exp_mw_5_to_10_mw_1_to_2 exp_mw_5_to_10_mw_20_above exp_mw_5_to_10_mw_2_to_3
#> 1              0.0008445436               1.508114e-04             0.0003921095
#> 2              0.0005899366               4.129556e-04             0.0007079240
#> 3              0.0008303402               4.744801e-04             0.0009885003
#> 4              0.0010051487               4.370212e-04             0.0006992339
#> 5              0.0010310628               4.909823e-05             0.0007037413
#> 6              0.0013422670               0.000000e+00             0.0006372379
#> 7              0.0012155287               0.000000e+00             0.0007012665
#> 8              0.0009614034               0.000000e+00             0.0002884210
#> 9              0.0008636297               1.919177e-05             0.0006333284
#> 10             0.0007633224               0.000000e+00             0.0006679071
#>    exp_mw_5_to_10_mw_3_to_5 iso_mw_5_to_10_mw_5_to_10
#> 1              0.0008445436              0.0006334077
#> 2              0.0010028923              0.0021827656
#> 3              0.0022933207              0.0017002205
#> 4              0.0015732763              0.0018791911
#> 5              0.0006382770              0.0003436876
#> 6              0.0004609806              0.0002982816
#> 7              0.0006077643              0.0005376377
#> 8              0.0003244736              0.0001562280
#> 9              0.0006525202              0.0003454519
#> 10             0.0003100997              0.0003339535
#>                              geom
#> 1  POLYGON ((608174.8 7542852,...
#> 2  POLYGON ((608355.9 7542806,...
#> 3  POLYGON ((608020.1 7542689,...
#> 4  POLYGON ((608089.5 7542497,...
#> 5  POLYGON ((607678.5 7542456,...
#> 6  POLYGON ((607427.8 7543110,...
#> 7  POLYGON ((607739.6 7543753,...
#> 8  POLYGON ((607956.1 7543728,...
#> 9  POLYGON ((608306.5 7543240,...
#> 10 POLYGON ((608688.1 7543833,...