Calculate segregation metrics

Calculate segregation metrics

measure_segregation(data, bandwidths = 0)

Arguments

data
a spatial sf object of the study area divided into areal units (such as census tracts), containing the columns 'id', 'geometry', and one column per population group indicating that group's population in the local areal unit.

bandwidth
numeric. A bandwidth, in meters, used to set the scale of analysis for spatial segregation measurement. When set to 0, aspatial metrics will be calculated.

Value

a segreg object, a list containing the input spatial data and the results of the segregation metrics.

Examples

library(sf)
library(segregr)

# load sample data from package segregr
marilia_sf <- st_read(system.file("extdata/marilia_2010.gpkg", package = "segregr"))

#> Reading layer `marilia_2010_proj' from data source `/private/var/folders/p_/v724hrfn46v1sxqm4xtxfv580000gn/T/RtmpSh0xbm/temp_libpatha1532b97fd4/segregr/extdata/marilia_2010.gpkg' using driver `GPKG'
#> Simple feature collection with 295 features and 10 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 597725.2 ymin: 7535553 xmax: 612714.8 ymax: 7551388
#> projected CRS:  SIRGAS 2000 / UTM zone 22S

# calculate segregation metrics
segregation <- measure_segregation(marilia_sf)

# global dissimilarity index
segregation$D

#>    bw         D
#> 1:  0 0.2749453

# global entropy
segregation$E

#> [1] 1.733732

# global information theory index H
segregation$H

#>    bw        H
#> 1:  0 0.110396